refactor(dashboard): split command-normalizers-swarm into domain sub-modules
Split command-normalizers-swarm.ts (971 LOC) into normalizers/swarm-*.ts modules.
Grouped by domain: lane, proof, worker, chain, help, orchestra, snapshot.
Original file becomes a barrel re-export (8 lines).
